sync(a5): SPMD multi-block dispatch, profiling fixes, and tensor dependency tracking (#426)
Synchronize A5 platform, runtimes, and tests with a2a3 improvements. Follows the established sync pattern.

Platform (src/a5/platform/):
- 012675a ([[[[[[#419](https://github.com/ChaoZheng109/simpler/issues/419)](https://github.com/ChaoZheng109/simpler/issues/419)](https://github.com/ChaoZheng109/simpler/issues/419)](https://github.com/ChaoZheng109/simpler/issues/419)](https://github.com/ChaoZheng109/simpler/issues/419)](https://github.com/ChaoZheng109/simpler/issues/419)): Add PerfSetDeviceCallback for device context setup in mgmt_loop, buffer recycling improvements
- fe63325 ([[[[[[#403](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)): Fix include paths in perf_profiling.h, rename params_cycle to args_cycle

Runtime host_build_graph (src/a5/runtime/host_build_graph/):
- 012675a ([[[[[[#419](https://github.com/ChaoZheng109/simpler/issues/419)](https://github.com/ChaoZheng109/simpler/issues/419)](https://github.com/ChaoZheng109/simpler/issues/419)](https://github.com/ChaoZheng109/simpler/issues/419)](https://github.com/ChaoZheng109/simpler/issues/419)](https://github.com/ChaoZheng109/simpler/issues/419)): Add implicit task profiling records in Case 1/Case 2, reinterpret_cast cleanup, license header

Runtime tensormap_and_ringbuffer (src/a5/runtime/tensormap_and_ringbuffer/):
- 7059fff ([[[[[[#389](https://github.com/ChaoZheng109/simpler/issues/389)](https://github.com/ChaoZheng109/simpler/issues/389)](https://github.com/ChaoZheng109/simpler/issues/389)](https://github.com/ChaoZheng109/simpler/issues/389)](https://github.com/ChaoZheng109/simpler/issues/389)](https://github.com/ChaoZheng109/simpler/issues/389)): Encapsulate TaskOutputTensors materialization in orchestrator
- 27a85c8 ([[[[[[#390](https://github.com/ChaoZheng109/simpler/issues/390)](https://github.com/ChaoZheng109/simpler/issues/390)](https://github.com/ChaoZheng109/simpler/issues/390)](https://github.com/ChaoZheng109/simpler/issues/390)](https://github.com/ChaoZheng109/simpler/issues/390)](https://github.com/ChaoZheng109/simpler/issues/390)): Const-qualify set_tensor_data Tensor parameter
- 1d97ac5 ([[[[[[#395](https://github.com/ChaoZheng109/simpler/issues/395)](https://github.com/ChaoZheng109/simpler/issues/395)](https://github.com/ChaoZheng109/simpler/issues/395)](https://github.com/ChaoZheng109/simpler/issues/395)](https://github.com/ChaoZheng109/simpler/issues/395)](https://github.com/ChaoZheng109/simpler/issues/395)): Arg inherits TaskArgs, simplify orchestrator arg passing
- 121a1d5 ([[[[[[#387](https://github.com/ChaoZheng109/simpler/issues/387)](https://github.com/ChaoZheng109/simpler/issues/387)](https://github.com/ChaoZheng109/simpler/issues/387)](https://github.com/ChaoZheng109/simpler/issues/387)](https://github.com/ChaoZheng109/simpler/issues/387)](https://github.com/ChaoZheng109/simpler/issues/387)): Add to_u64/from_u64 type-safe conversion utilities in orchestration API
- fe63325 ([[[[[[#403](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)): Defer output tensor materialization, TensorCreateInfo pointer, tensormap link_entry
- cd59b47 ([[[[[[#404](https://github.com/ChaoZheng109/simpler/issues/404)](https://github.com/ChaoZheng109/simpler/issues/404)](https://github.com/ChaoZheng109/simpler/issues/404)](https://github.com/ChaoZheng109/simpler/issues/404)](https://github.com/ChaoZheng109/simpler/issues/404)](https://github.com/ChaoZheng109/simpler/issues/404)): Add SPMD context accessors, intrinsic.h, build_payload with LocalContext/GlobalContext
- 4917d12 ([[[[[[#415](https://github.com/ChaoZheng109/simpler/issues/415)](https://github.com/ChaoZheng109/simpler/issues/415)](https://github.com/ChaoZheng109/simpler/issues/415)](https://github.com/ChaoZheng109/simpler/issues/415)](https://github.com/ChaoZheng109/simpler/issues/415)](https://github.com/ChaoZheng109/simpler/issues/415)): Refine tensor dependency tracking, TensorCreateInfo alignment, owner tracking
- 34a6e1c ([[[[[[#417](https://github.com/ChaoZheng109/simpler/issues/417)](https://github.com/ChaoZheng109/simpler/issues/417)](https://github.com/ChaoZheng109/simpler/issues/417)](https://github.com/ChaoZheng109/simpler/issues/417)](https://github.com/ChaoZheng109/simpler/issues/417)](https://github.com/ChaoZheng109/simpler/issues/417)): SPMD multi-block dispatch, scheduler dual-queue, submit_types extensions

Tests (examples/a5/, tests/st/a5/):
- be765f1 ([[[[[[#392](https://github.com/ChaoZheng109/simpler/issues/392)](https://github.com/ChaoZheng109/simpler/issues/392)](https://github.com/ChaoZheng109/simpler/issues/392)](https://github.com/ChaoZheng109/simpler/issues/392)](https://github.com/ChaoZheng109/simpler/issues/392)](https://github.com/ChaoZheng109/simpler/issues/392)): Migrate paged_attention orchestration to ChipStorageTaskArgs API
- 121a1d5 ([[[[[[#387](https://github.com/ChaoZheng109/simpler/issues/387)](https://github.com/ChaoZheng109/simpler/issues/387)](https://github.com/ChaoZheng109/simpler/issues/387)](https://github.com/ChaoZheng109/simpler/issues/387)](https://github.com/ChaoZheng109/simpler/issues/387)](https://github.com/ChaoZheng109/simpler/issues/387)): Use from_u64<float> in softmax_prepare kernels
- fe63325 ([[[[[[#403](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)](https://github.com/ChaoZheng109/simpler/issues/403)): Add license headers, NOLINT annotations, output tensor view(true)
- cd59b47 ([[[[[[#404](https://github.com/ChaoZheng109/simpler/issues/404)](https://github.com/ChaoZheng109/simpler/issues/404)](https://github.com/ChaoZheng109/simpler/issues/404)](https://github.com/ChaoZheng109/simpler/issues/404)](https://github.com/ChaoZheng109/simpler/issues/404)](https://github.com/ChaoZheng109/simpler/issues/404)): Add spmd_basic example (AIC+AIV SPMD read test)
- 34a6e1c ([[[[[[#417](https://github.com/ChaoZheng109/simpler/issues/417)](https://github.com/ChaoZheng109/simpler/issues/417)](https://github.com/ChaoZheng109/simpler/issues/417)](https://github.com/ChaoZheng109/simpler/issues/417)](https://github.com/ChaoZheng109/simpler/issues/417)](https://github.com/ChaoZheng109/simpler/issues/417)): Add spmd_multiblock_aiv and spmd_multiblock_mix examples
- Remove redundant end-of-kernel sync barriers in paged_attention test kernels
- Adjust paged_attention orch_thread_num (2→1), paged_attention_unroll block_dim (36→24)

+++ tests/st/a5/host_build_graph/paged_attention/kernels/orchestration/paged_attention_orch.cpp
@@ -173,19 +173,14 @@ int build_paged_attention_graph(Runtime *runtime, const ChipStorageTaskArgs &orc
 
             for (uint32_t bn = 0; bn < bn_this_batch; bn++) {
                 int cur_block_idx = host_block_table[b_idx * max_num_blocks + bn];
-                int valid_len = std::min(
-                    static_cast<int>(block_size), cur_seq - static_cast<int>(bn) * static_cast<int>(block_size)
-                );
+                int valid_len = std::min(static_cast<int>(block_size), cur_seq - static_cast<int>(bn * block_size));
 
                 // Key: (total_blocks, block_size, kv_head_num, head_dim) bf16
-                // Stride to block: cur_block_idx * (block_size * kv_head_num * head_dim)
-                // Then offset to kv_head: kv_head_idx * head_dim (within each token row)
-                // But since we want contiguous (block_size, head_dim), and kv_head_num=1 makes it simple:
                 uint8_t *kj_ptr = reinterpret_cast<uint8_t *>(dev_key_cache) +
                                   (static_cast<int64_t>(cur_block_idx) * block_size * kv_head_num + kv_head_idx) *
                                       head_dim * sizeof(uint16_t);
 
-                // Value: (total_blocks, block_size, kv_head_num, head_dim) bf16 - same layout as key
+                // Value: (total_blocks, block_size, kv_head_num, head_dim) bf16
                 uint8_t *vj_ptr = reinterpret_cast<uint8_t *>(dev_value_cache) +
                                   (static_cast<int64_t>(cur_block_idx) * block_size * kv_head_num + kv_head_idx) *
                                       head_dim * sizeof(uint16_t);
